Fixed setup; Multiple interfaces now allowed
The inputs are now correctly initialized; previously it was possible
that the digital inputs appeared to be all HIGH because the setup code
didn't wait for the analog timers to expire.
Also added code to make it possible to connect multiple interfaces via
the expansion connector on the 30520. Unfortunately I only have one
interface so I couldn't test it, but the code to do this is pretty
trivial.