Some reorganization
Added the number of interfaces to the class. Also removed the number of interfaces parameters from the get/set functions and added such a parameter to the constructor and setup code.
Added some inline functions with the old parameters for backwards compatibility.
Renamed the Setup function to reduce confusion with the main setup() function in Arduino programs.
Some minor changes to comments.