Adding classes for manager, input pins, output pins and motors
The FishduinoMgr class stores the current state of the input and output pins, and provides functions to synchronize the data with the interface(s). This makes it unnecessary for the main program to keep track of the individual input and output pins of all interfaces.
The FishduinoInPin, FishduinoOutPin and FishduinoMotor classes abstract one or more input or output pins, and allow the program to manipulate those pins with functions that are almost like human language.

I will check in some code to my FischerTechnik Clock repo to demonstrate this, next.