Added Fishduino_Clock sample program to Fishduino library
Instead of starting a separate project as I implied in the previous commit, I added Fishduino_Clock to the Fishduino project as a sample. This is not complete yet, it simply forwards the minute hand once a minute.
There is some unused data in this, which will be used later.
There's also a minor problem with the current code: every once in a while the motor stops too soon because when the switch turns from off->on sometimes there is some bouncing and every once in a while the bouncing is detected as the switch going on, then off again.