Added a function to the Clock program to let motor stop at a home position
In the previous commit, I simply tested to see if the minute sensor had gone from ON to OFF, and that caused a problem because of a bouncing switch. By turning the motor on whenever the switch is on, that problem was fixed. But then I realized that I'll need the same functionality for the adjustment part of the Fishduino clock, so I created a generic function in the clock program to take care of this, and modified the input pin and motor modules to accommodate.